Change Ids from Long to Int for memory efficiency.

+++ Cchbc.Features/DashboardModule/Objects/DbFeatureVersionRow.cs
@@ -4,10 +4,10 @@ namespace Cchbc.Features.DashboardModule.Objects
 {
 	public sealed class DbFeatureVersionRow
 	{
-		public readonly long Id;
+		public readonly int Id;
 		public readonly string Name;
 
-		public DbFeatureVersionRow(long id, string name)
+		public DbFeatureVersionRow(int id, string name)
 		{
 			if (name == null) throw new ArgumentNullException(nameof(name));
