+++ Cchbc/Features/Data/DbFeatureContextRow.cs
@@ -4,10 +4,10 @@ namespace Cchbc.Features.Data
 {
 	public sealed class DbFeatureContextRow
 	{
-		public readonly long Id;
+		public readonly int Id;
 		public readonly string Name;
 
-		public DbFeatureContextRow(long id, string name)
+		public DbFeatureContextRow(int id, string name)
 		{
 			if (name == null) throw new ArgumentNullException(nameof(name));
