+++ Cchbc/Features/Data/DbFeatureExceptionRow.cs
@@ -4,10 +4,10 @@ namespace Cchbc.Features.Data
 {
 	public sealed class DbFeatureExceptionRow
 	{
-		public readonly long Id;
+		public readonly int Id;
 		public readonly string Contents;
 
-		public DbFeatureExceptionRow(long id, string contents)
+		public DbFeatureExceptionRow(int id, string contents)
 		{
 			if (contents == null) throw new ArgumentNullException(nameof(contents));
