+++ Cchbc/Features/Data/DbFeatureStepRow.cs
@@ -4,10 +4,10 @@ namespace Cchbc.Features.Data
 {
 	public sealed class DbFeatureStepRow
 	{
-		public readonly long Id;
+		public readonly int Id;
 		public readonly string Name;
 
-		public DbFeatureStepRow(long id, string name)
+		public DbFeatureStepRow(int id, string name)
 		{
 			if (name == null) throw new ArgumentNullException(nameof(name));
